backend/requirements.txt
aiofiles==25.1.0
aiosmtplib==4.0.2
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.11.0
axios==0.4.0
babel==2.17.0
backrefs==5.9
bcrypt==4.2.1
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
dotenv==0.9.9
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
fakeredis==2.32.0
fastapi==0.119.1
fastapi-mail==1.5.7
flake8==7.3.0
ghp-import==2.1.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
isort==7.0.0
Jinja2==3.1.6
jwt==1.4.0
lxml==6.0.2
Mako==1.3.10
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.23
mkdocs-material-extensions==1.3.1
numpy==2.3.3
openpyxl==3.1.2
packaging==25.0
paginate==0.5.7
pandas==2.3.3
passlib[bcrypt]==1.7.4
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pyarrow==21.0.0
pyasn1==0.6.1
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pylance==0.36.0
pymdown-extensions==10.16.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
pyyaml_env_tag==1.1
redis==7.0.1
regex==2025.10.23
reportlab==4.4.3
requests==2.32.5
rich==14.2.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.37.0
watchdog==6.0.0
playwright>=1.40.0
